Extract getting started into separate component
